12_OpenAI_Agents_SDK/02_hello_agent/chat_history_stateful/pyproject.toml
[project]
name = "stateful-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.8.0",
    "openai-agents>=0.3.0",
]
